~ IMPORTANT COMMIT: Implemented the functionality for the Edit-Button in all Status-Cards by creating a base view model for orchestrating the Edit-Option and new view models for the active and passive Care-Statuses that deal with the Edit-Option. Also added a new view for the Edit-Option, in which the user can make changes on existing values of the Care-Statuses. To make this possible, a new Command for the base view model of all Care-Statuses (CareStatusViewModel.cs) was added and and integrated in the view models of all Care-Statuses as well as bound to the DataTemplate in the Status-Card-Styles (StatusCardStyles.xaml)

+++ GreenKeeper/Models/SunlightRequirement.cs
@@ -11,7 +11,7 @@ namespace GreenKeeper.Models
     {
         public int Id { get; set; }
         public int PlantId { get; set; }
-        public double Hours { get; set; }
+        public int Hours { get; set; }
         public SunlightPeriod Period { get; set; }
     }
 }
